requirements.txt
Django==4.2.6
djangorestframework==3.14.0
djoser==2.2.0
psycopg2==2.9.9
